# lib/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MarkPad API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Project {
  id: number;
  name: string;
  template: string;
  createdAt: string;
  updatedAt: string;
}

export type CreateProjectBodyTemplate =
  (typeof CreateProjectBodyTemplate)[keyof typeof CreateProjectBodyTemplate];

export const CreateProjectBodyTemplate = {
  plain: "plain",
  academic: "academic",
  report: "report",
  letter: "letter",
} as const;

export interface CreateProjectBody {
  name: string;
  template: CreateProjectBodyTemplate;
}

export interface UpdateProjectBody {
  name: string;
}

export type FileEntryType = (typeof FileEntryType)[keyof typeof FileEntryType];

export const FileEntryType = {
  file: "file",
  directory: "directory",
} as const;

export interface FileEntry {
  path: string;
  name: string;
  type: FileEntryType;
  size?: number;
}

export interface FileContent {
  path: string;
  content: string;
  wordCount: number;
  lastSavedAt: string;
}

export interface CreateFileBody {
  path: string;
  content?: string;
}

export interface SaveFileBody {
  content: string;
}

export interface Snapshot {
  id: number;
  filePath: string;
  wordCount: number;
  /** @nullable */
  label?: string | null;
  createdAt: string;
}

export interface HistoryResponse {
  pinned: Snapshot[];
  today: Snapshot[];
  yesterday: Snapshot[];
  older: Snapshot[];
}

export interface PinSnapshotBody {
  label: string;
}

export interface RenderBody {
  content: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface ProjectStats {
  totalFiles: number;
  totalWords: number;
  totalSnapshots: number;
  lastEditedAt: string;
}

export interface Template {
  id: string;
  name: string;
  description: string;
  content: string;
}

export type GetHistoryParams = {
  /**
   * File path to get history for
   */
  file: string;
};

export type ExportProjectParams = {
  /**
   * Export format
   */
  format: ExportProjectFormat;
  /**
   * File to export (defaults to main.md)
   */
  file?: string;
};

export type ExportProjectFormat =
  (typeof ExportProjectFormat)[keyof typeof ExportProjectFormat];

export const ExportProjectFormat = {
  pdf: "pdf",
  md: "md",
  latex: "latex",
} as const;