# lib/lib/api-zod/src/generated/types/createProjectBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MarkPad API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateProjectBodyTemplate } from "./createProjectBodyTemplate";

export interface CreateProjectBody {
  name: string;
  template: CreateProjectBodyTemplate;
}