# lib/lib/api-zod/src/generated/types/fileContent.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MarkPad API specification
 * OpenAPI spec version: 0.1.0
 */

export interface FileContent {
  path: string;
  content: string;
  wordCount: number;
  lastSavedAt: string;
}