# lib/lib/api-zod/src/generated/types/fileEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MarkPad API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FileEntryType } from "./fileEntryType";

export interface FileEntry {
  path: string;
  name: string;
  type: FileEntryType;
  size?: number;
}