# lib/lib/api-zod/src/generated/types/historyResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MarkPad API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Snapshot } from "./snapshot";

export interface HistoryResponse {
  pinned: Snapshot[];
  today: Snapshot[];
  yesterday: Snapshot[];
  older: Snapshot[];
}