# lib/lib/api-zod/src/generated/types/template.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MarkPad API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Template {
  id: string;
  name: string;
  description: string;
  content: string;
}